requirements.txt
psutil==5.4.5
pyserial==3.4
